Refactoring and unit testing

+++ src/main/java/com/master/BioskopVozdovac/actor/service/ActorService.java
@@ -4,10 +4,10 @@ import com.master.BioskopVozdovac.actor.adapter.ActorAdapter;
 import com.master.BioskopVozdovac.actor.model.ActorDTO;
 import com.master.BioskopVozdovac.actor.model.ActorEntity;
 import com.master.BioskopVozdovac.actor.repository.ActorRepository;
-import com.master.BioskopVozdovac.exception.UserException;
+import com.master.BioskopVozdovac.exception.NotFoundException;
 import lombok.RequiredArgsConstructor;
-import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
 
 import java.util.List;
 
@@ -18,6 +18,7 @@ import java.util.List;
  */
 @Service
 @RequiredArgsConstructor
+@Transactional(readOnly = true)
 public class ActorService {
 
     /**
@@ -26,20 +27,16 @@ public class ActorService {
     private final ActorRepository actorRepository;
 
     /**
-     * Adapter for converting between DTOs and entities
-     */
-    private final ActorAdapter actorAdapter;
-
-    /**
      * Saves a new actor based on the provided ActorDTO.
      *
      * @param dto The ActorDTO containing actor details to be saved.
      *
      * @return The saved ActorDTO after converting from the saved ActorEntity.
      */
+    @Transactional
     public ActorDTO saveActor(ActorDTO dto) {
-        ActorEntity entity = actorRepository.saveAndFlush(actorAdapter.dtoToEntity(dto));
-        return actorAdapter.entityToDTO(entity);
+        ActorEntity entity = actorRepository.save(ActorAdapter.dtoToEntity(dto));
+        return ActorAdapter.entityToDTO(entity);
     }
 
     /**
@@ -47,12 +44,12 @@ public class ActorService {
      *
      * @param id The ID of the actor to retrieve.
      * @return The ActorDTO representing the actor found by the ID.
-     * @throws UserException if no actor is found with the specified ID.
+     * @throws NotFoundException if no actor is found with the specified ID.
      */
     public ActorDTO getActorById(Long id) {
         ActorEntity entity = actorRepository.findById(id).orElseThrow(()
-                -> new UserException("Could not find actor with id: " + id, HttpStatus.NOT_FOUND));
-        return actorAdapter.entityToDTO(entity);
+                -> new NotFoundException("Could not find actor with id: " + id));
+        return ActorAdapter.entityToDTO(entity);
     }
 
     /**
@@ -61,9 +58,21 @@ public class ActorService {
      * @param dto The ActorDTO containing updated actor details.
      * @return The updated ActorDTO after converting from the updated ActorEntity.
      */
+    @Transactional
     public ActorDTO updateActor(ActorDTO dto) {
-        ActorEntity entity = actorRepository.saveAndFlush(actorAdapter.dtoToEntity(dto));
-        return actorAdapter.entityToDTO(entity);
+        ActorEntity entity = actorRepository.findById(dto.actorID())
+                .orElseThrow(() -> new NotFoundException("Actor not found"));
+
+        updateActorEntity(entity, dto);
+
+        ActorEntity savedEntity = actorRepository.save(entity);
+        return ActorAdapter.entityToDTO(savedEntity);
+    }
+
+    private void updateActorEntity(ActorEntity entity, ActorDTO dto) {
+        entity.setFirstName(dto.firstName());
+        entity.setLastName(dto.lastName());
+        entity.setGender(dto.gender());
     }
 
     /**
@@ -72,7 +81,11 @@ public class ActorService {
      * @param id The ID of the actor to delete.
      * @return A success message indicating the deletion of the actor with the specified ID.
      */
+    @Transactional
     public String deleteActorById(Long id) {
+        if (!actorRepository.existsById(id))
+            throw new NotFoundException("Actor not found");
+
         actorRepository.deleteById(id);
         return "Successfully deleted actor with id: " + id;
     }
@@ -84,6 +97,6 @@ public class ActorService {
      * @return List of ActorDTO objects representing all actors in the database.
      */
     public List<ActorDTO> getAll() {
-        return actorAdapter.toDTO(actorRepository.findAll());
+        return ActorAdapter.toDTO(actorRepository.findAll());
     }
 }
